Extract decorator types into own file
